Centralize more formatting
